## Description

Add partitionId for history cleanup relevant tables. This is needed to
do cleanup per partition, so that multiple exporters (one per partition)
can cleanup there "own" data.

+++ zeebe/exporters/rdbms-exporter/src/main/java/io/camunda/exporter/rdbms/RdbmsExporter.java
@@ -29,7 +29,7 @@ public class RdbmsExporter {
   private final Map<ValueType, List<RdbmsExportHandler>> registeredHandlers;
   private Controller controller;
 
-  private final long partitionId;
+  private final int partitionId;
   private final RdbmsWriter rdbmsWriter;
 
   // configuration
